Workaround flatbuffer returning invalid buffer sequence
I might be misunderstanding something but basic_flat_buffer::data
should be return a ConstBufferSequence type. However, whenever I
try to iterate over it, it appears to be giving garbage on the
second iteration. We know flatbuffer will always have exactly one
buffer so this isn't the biggest deal ever but a tad annoying.